requirements.txt
flask
flask_socketio
python-socketio
eventlet
openai
pillow
python-dotenv
httpx

# Core dependencies for running the 5th edition AI dungeon master system

# Google Gemini API for AI model interactions (migrated from OpenAI)
# openai>=1.0.0  # Removed - migrated to Gemini

# Google Generative AI for Gemini integration
google-generativeai>=0.8.0

# Web interface framework
Flask>=2.3.0
Flask-SocketIO>=5.3.0
waitress>=2.1.2
python-socketio>=5.9.0

# JSON schema validation for game data
jsonschema>=4.17.0

# Terminal output coloring
termcolor>=2.3.0

# Progress bars for long operations
tqdm>=4.65.0

# HTTP requests for web interactions
requests>=2.31.0

# Google Drive API for Cloud Storage
google-api-python-client>=2.88.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Token counting (legacy from OpenAI - may need adjustment for Gemini)
tiktoken>=0.5.0

# Python standard library extensions (usually included)
# pathlib - included in Python 3.4+
# datetime - standard library
# json - standard library
# os - standard library
# sys - standard library
# re - standard library
# subprocess - standard library
# threading - standard library
# queue - standard library
# time - standard library
# io - standard library
# contextlib - standard library
# shutil - standard library
# codecs - standard library

# Optional: For enhanced development experience
# black>=23.0.0          # Code formatting
# flake8>=6.0.0          # Code linting
# pytest>=7.0.0          # Testing framework
# mypy>=1.0.0            # Type checking

# Note: Make sure you have Python 3.9+ installed
# Install with: pip install -r requirements.txt